Bot

Description

IRC bot interface. Implements some IRC-specific logic for the bot, and re-exports everything needed to program the bot.

Bot-specific keys for global store

currChanList :: GlobalKey [Channel]

Channels the bot is currently in.

autoJoinList :: PersistentKey [Channel]

Channels the bot should autojoin.

nameCallbacks :: GlobalKey [(String, [User] -> Bot ())]

List of callbacks for active NAME queries, to be invoked when the server replies.

whoisCallbacks :: GlobalKey [(String, Maybe User -> Bot ())]

List of callbacks for active WHOIS/WHOWAS queries, to be invoked when the server replies.

statusChars :: PersistentKey [Char]

Character prefixes for users that indicate status in channels; e.g., '@' for op or + for voice.

statusOpChars :: PersistentKey [Char]

Status prefixes that indicate op privileges.

IO

Messaging

sendMessage :: RecipientC a => a -> String -> Bot ()

Send a textual message to the recipient

Queries

getWhois :: User -> (Maybe User -> Bot ()) -> Bot ()

Make a WHOIS query for the given user, invoking the callback when the server replies.

getWhowas :: User -> (Maybe User -> Bot ()) -> Bot ()

Make a WHOWAS query for the given user, invoking the callback when the server replies.

getNames :: Channel -> ([User] -> Bot ()) -> Bot ()

Make a NAME query (list of nicks in the channel) for the given channel, invoking the callback when the server replies.

getOwnStatus :: Channel -> (Maybe Char -> Bot ()) -> Bot ()

Find the bot's own status character (or Nothing if it doesn't have any) in the given channel, invoking the callback when it finds it.

Channel management

joinChannels :: [Channel] -> Bot ()

Join the list of channels, where each element is a channel name string (e.g., "#channame")

kickUser :: Channel -> User -> Bot ()

Kick the given user from the given channel, with no reason.

kickUserFor :: Channel -> User -> String -> Bot ()

Kick the given user from the given channel, with the reason given.

banMask :: Channel -> String -> Bot ()

Ban the given mask from the given channel.

unbanMask :: Channel -> String -> Bot ()

Unban the given mask from the given channel.

Re-exported

module Bot.Bot

module Bot.Msg

module Bot.Store