Bot.Msg

Description

Defines the internal message/event structure, and implements parsing/serialising.

Server events are either messages that the bot has received from the server, like a PRIVMSG, or other events like the bot starting up. They're parsed from the textual line received by a parser using Parsec, and for common commands with specific syntax, they're parsed in a detailed way.

Client messages are messages that will be sent to the server by the bot, usually within a callback. Client messages can be sent generically, as a stock CMsg.

Message structure

Generic

data User

A user with a nickname, and maybe user/host/other things. This is multipurpose, used just to indicate that some nick is a user (when used with makeUser), or, e.g., as a result from a NAMES or WHOIS request containing more info about the user.

Constructors

User

Fields

nick :: String

Nick

user :: Maybe String

User, if we know it.

host :: Maybe String

Host, if we know it.

statusCharL :: [(Channel, Char)]

A list of status prefixes that the user has in each channel that we know (e.g., @ for op).

newtype Channel

A channel; the string doesn't include the # prefix.

Constructors

Channel String

newtype Server

A server and its hostname.

Constructors

Server String

class SenderC a where

Sender class: User and Server can both be senders in a message (see Sender).

Minimal complete definition

fromS

Methods

fromS :: Sender -> a

Sender to User or Server

class RecipientC a where

Recipient class: User and Channel can both be recipients in a message (see Recipient).

Minimal complete definition

fromR, toR

Methods

fromR :: Recipient -> a

Recipient to User or Channel

toR :: a -> Recipient

User or Channel to Recipient

fromS :: SenderC a => Sender -> a

Sender to User or Server

fromR :: RecipientC a => Recipient -> a

Recipient to User or Channel

toR :: RecipientC a => a -> Recipient

User or Channel to Recipient

makeUser :: String -> User

Make a user struct from just a nick.

makeUserH :: String -> String -> String -> User

Make a user struct from a nick, user string and host string.

makeChannel :: [Char] -> Channel

Take a channel name of the form "#chan" and create a Channel.

channelToStr :: Channel -> [Char]

Turn a Channel into a channel name of the form "#chan".

Event

data SEvent

Events: a message that's been sent by the server and received by the client, or an event initiated on the bot side like startup. The messages have detailed structures; they're parsed into easily usable forms for callbacks. The exception is SNumeric, which is generic for all numeric messages and may have any number of arguments.

Constructors

SPing

Fields

srv :: String

SNotice

Fields

from :: Sender

to :: Recipient

text :: String

SPrivmsg

Fields

from :: Sender

to :: Recipient

text :: String

SNumeric

Fields

fromMaybe :: Maybe Sender

n :: Int

args :: [String]

SJoin

Fields

from :: Sender

to :: Recipient

SPart

Fields

from :: Sender

to :: Recipient

SKick

Fields

from :: Sender

to :: Recipient

target :: String

reason :: String

SMode

Fields

from :: Sender

modeTarget :: Recipient

modeChanges :: String

Startup

Connected

data Sender

Sender field - can be either user or server.

Constructors

SUser String String String
nick, user, host

SServer String
host

data Recipient

Recipient field - can either be user or channel.

Constructors

RUser String

RChannel String

Client message

data CMsg

Message that will be sent by the bot. The structure is fairly barebones because it's not going to be parsed; whatever is sending the message can implement the structure correctly itself, or a helper function in Bot can handle that.

Constructors

CMsg

Fields

command :: ClientCmd

argsC :: [String]

data ClientCmd

Commands that will be sent by the bot; deliberately named identically to their textual representations, to make serialising easy.

Constructors

NICK

USER

JOIN

PONG

PRIVMSG

NAMES

WHOIS

WHOWAS

KICK

MODE

Parsing and serialising

joinMsg :: CMsg -> String

Convert a message to a string in the IRC format.

readMsg :: String -> Either ParseError SEvent

Parse a message string in the IRC format.

Parsing internals

parseMsg :: Parser SEvent

The main parser for messages.

parseSender :: Parser Sender

Parse the sender field of a message, which will be a :, followed by either a user in the form nick!user@host, or a server in the form host.

parseRecipient :: Parser Recipient

Parse a recipient field, which might either be a channel in the form #channel-name, or a user in the form nick.

parseArgs :: Parser [String]

Parse the arguments at the end of the message.

parseUntil :: String -> Parser String

Parse until one of the characters in s, or eof.

parseWord :: Parser String

Parse until one of space : r n eof, and eat a trailing space if it's there.