Bot.Msg.Splices

Documentation

makeNMsg :: Integer -> Q Exp

A splice for generating a function that helps parse a message of the format :source command target [args] where there are n args. The function takes a constructor, which is applied to the source, target, and the arguments. For example:

$(makeNMsg 2) :: (Sender -> Recipient -> String -> String -> SEvent) -- ^ The constructor, taking two additional arguments
                 -> Maybe Sender -- ^ The sender, if present (if not, a parse error will result).
                 -> [String] -- ^ The arguments in the message
                 -> Parser SEvent -- ^ Resulting parser, returning the constructed SEvent