Bot.Scripting.Core

Description

Implements the core IRC-related functionality, like user registration, responding to pings, joining and leaving channels, and so on.

Documentation

callbacks :: [SEvent -> Bot ()]

Event hooks

Connection

connected :: SEvent -> Bot ()

User registration, and set a callback to join channels.

respondToEndOfMOTD :: SEvent -> Bot ()

Join channels on 376 (end of MOTD), and then unhook this callback.

Queries

respondToNAMES :: SEvent -> Bot ()

NAMES reply - handles both RPL_NAMREPLY (353) and RPL_ENDOFNAMES (366); the requester will be called on 353 and unhooked; or if, e.g., the channel doesn't exist, the server will only send a 366 and the requester will be called on that.

respondToWHOISUSER :: SEvent -> Bot ()

WHOIS reply - like respondToNAMES, this handles both RPL_WHOISUSER (311) and RPL_ENDOFWHOIS (318). It also handles the equivalent WHOWAS responses, since they're virtually identical - RPL_WHOWASUSER (314) and RPL_ENDOFWHOWAS (369).

namesResponse :: String -> [User] -> Bot ()

Find the requester for the NAMES query, unhook and invoke it.

whoisResponse :: String -> Maybe User -> Bot ()

Find the requester for the WHOIS/WHOWAS query, unhook and invoke it.

Channel management

respondToJoin :: SEvent -> Bot ()

If it's the bot that's joined, update the current channels list.

respondToPart :: SEvent -> Bot ()

If it's the bot that's parted, update the current channels list.

respondToKick :: SEvent -> Bot ()

If it's the bot that's been kicked, update the current channels list.

Other

respondToPing :: SEvent -> Bot ()

Respond to a server PING with PONG.