Bot.Scripting.Flood

Description

Flood control. There are four parameters: timeWindow, linesWeight, charsWeight, joinWeight. The bot will kick anyone who, in a single channel, posts enough messages within timeWindow seconds such that number-of-lines * linesWeight + number-of-characters * charsWeight > 1.

e.g., if you want the bot to kick anyone who sends more then 6 lines in a 5 second window, set timeWindow = 5, linesWeight = 0.1667 (1/6), charsWeight = 0.0.

The bot will only do this in channels. This also handles join flooding, by treating a join message as equivalent to joinWeight zero-character messages. e.g., if joinWeight is 2 in the above example, someone who joins more than three times in five seconds would be kicked.

Settings

channels :: PersistentKey [Channel]

timeWindow :: PersistentKey Int

linesWeight :: PersistentKey Float

charsWeight :: PersistentKey Float

joinWeight :: PersistentKey Float

Flood control

type MessageMap = Map (Channel, User) [(UTCTime, (Float, Float))]

Message history is stored per user per channel, in a Map with keys (Channel, User). The values are lists of (time message was sent, (how many lines the message counts as, number of characters in message)) for each message sent by that user to that channel and received by the bot within timeWindow; although these values are in general floats.

trackMessage :: SEvent -> Bot ()

Main event hook. Captures various channel events and weights them accordingly, passing them on to updateUserMessage.

cleanLines :: (Channel, User) -> Bot ()

Remove any logs outside the time window from the user's message history.

addLine :: (Channel, User) -> (Float, Float) -> Bot ()

Add the given line to the user's message history.

isFlooding :: (Channel, User) -> Bot Bool

Check whether the user is flooding given their message history.

stopFlooding :: (Channel, User) -> Bot ()

Kick a user for flooding.

callbacks :: [SEvent -> Bot ()]

Event hooks