Bot.Scripting.Laws

Description

Auto-enforce rules like not allowing users to have certain characters in their nick.

Documentation

callbacks :: [SEvent -> Bot ()]

Event hooks

Forbid characters

Settings

channels :: PersistentKey [Channel]

Channels to operate in.

forbiddenChars :: PersistentKey [Char]

Characters to forbid users having in their nicks.

graceOffences :: PersistentKey Int

Number of prior offences to allow before instakicking the offender.

graceTime :: PersistentKey Int

Time to wait before kicking the offender, if they're in grace.

offendersL :: PersistentKey [(String, Int)]

List of hosts, and numbers of offences.

exceptionsL :: PersistentKey [String]

Exception nicks to allow.

warningMessage :: PersistentKey String

Message to send an offender when in grace.

Logic

listenNick :: SEvent -> Bot ()

Listen for a nick change event. These don't come associated with a channel, so we WHOIS lookup the user in question to find all the channels they're in. We then see whether there are any that are in our channels list, calling handleNickEv if so.

listenJoin :: SEvent -> Bot ()

Listen for a join event, checking whether the channel is in our channels list and calling handleNickEv if so.

handleNickEv :: Channel -> String -> String -> Bot ()

'handleNickEv ch newnick host': Handle a nick event in the given channel ch. The user with nick newnick and host host has changed nicks or joined the channel. Check if their nick contains a forbidden character, and if so, deal with it - look up the number of offences they've already made (by their host), warning them if they're under the threshold for instakick (graceOffences) and instakicking them if they're over.

grace :: Channel -> String -> Bot ()

Warn a user, and set a timer to kick them soon.

punish :: Channel -> String -> Bot ()

Kick a user.