Bot.Scripting.Vote

Documentation

callbacks :: [SEvent -> Bot ()]

Event hooks

voteChan :: PersistentKey Recipient

voteActive :: PersistentKey Bool

voteOpts :: PersistentKey VoteOpt

yesHosts :: PersistentKey [String]

noHosts :: PersistentKey [String]

commandChar :: Char

The commands a user can give in the channel.

commands :: [([Char], SEvent -> Bot ())]

acts :: [[Char]]

Actions that can be voted on, and their vote lengths in seconds.

actl :: [([Char], Integer)]

onStartup :: SEvent -> Bot ()

Clear the current vote (for dev purposes).

processCommand :: SEvent -> Bot ()

Listen to all channel messages, and extract those which are commands for this bot.

vote :: SEvent -> Bot ()

!vote - command to start a vote.

startVote :: VoteOpt -> Bot ()

Start the given vote.

data VoteChoice

!yes or !no - vote yes or no, checking the user's host to see if they've already voted.

Constructors

VYes

VNo

castVote :: VoteChoice -> SEvent -> Bot ()

voteInfo :: Monad m => t -> m ()

Return info about the current vote.

voteEnd :: Bot ()

End the current vote, taking the action if successful.

parseUntil' :: String -> ParsecT String () Identity String

parseCommand :: Parser (SEvent -> Bot ())

Parse the top-level command, returning an action to take.

data VoteOpt

Vote options. Not all possible combinations are valid votes - they may have Nothings for necessary fields.

Constructors

VoteOpt

Fields

actionO :: Maybe String

targetO :: Maybe String

lenO :: Maybe Int

reasonO :: Maybe String

parseVoteCmd :: Parser VoteOpt

Parse a !vote command, returning vote options.