Bot.Store

Description

A store mapping keys to values of arbitrary type. GlobalStore is a hashmap from (String, TypeRep) pairs to Dynamic objects; the keys used, GlobalKey, contain information about the type, so that the values can be retrieved from the Dynamic instances.

The keys use strings for uniqueness, but are also unique by type. So two different keys will only point to the same object if they have the same key string and default values of the same type.

The keys also contain a default value, which is returned whenever the lookup fails. This makes it seem as though the store was initialised with the default values for all the keys.

GlobalStore is intended to provide a scripting interface for addon modules, by providing a way for users to specify arbitrary stateful variables, and share them across modules by exporting the keys.

Documentation

type GlobalStore = HashMap (TypeRep, String) Dynamic

Store for keeping global state of arbitrary type. It has keys of type (TypeRep, String), allowing for uniqueness for keys of the same type and guaranteeing that keys of different types will never collide. It stores values as Dynamic, extracting them using the type information provided by the key.

empty :: GlobalStore

An empty GlobalStore.

data GlobalKey a

Key for GlobalStore - a is the type that's stored. GlobalKey a s creates a global key with:

default value a (which also specifies the type of the value being stored)

id string s (which should be unique for keys of the same type)

Constructors

GlobalKey a String

makeK :: Typeable a => GlobalKey a -> (TypeRep, String)

Convenience function for converting GlobalKey to '(TypeRep, String)'.

getGlobalFromStore :: Typeable a => GlobalStore -> GlobalKey a -> a

Look up the key in the store, returning the value found if there is one, or the default value contained in the key otherwise.

setGlobalToStore :: Typeable a => GlobalStore -> GlobalKey a -> a -> GlobalStore

Set the given value to the key in the store.

isInStore :: Typeable a => GlobalStore -> GlobalKey a -> Bool

Check whether the given key is in the given store.

class Typeable k a

Minimal complete definition

typeRep#