Main

Description

Top-level module for the bot. Deals with command-line arguments, loading the config, and initialising the bot state.

Command line options

data Flag

Command line flags

Constructors

Config String
-c configfile or --config configfile

Help
-h or --help

parseCmdArgs :: IO (Either String String)

Parse the command line arguments, returning Right config if there's a config file location specified, or Left err with an error message otherwise.

Config

loadConfig :: String -> IO (Either String GlobalStore)

Load a config file in the ini format from loc, returning Right st, a store initialised with the values found, if the config sets all the relevant keys in configKeys. If there are some missing keys or a problem with the config, return Left err instead, with an error message.

Main

main :: IO ()