Msg

Description

Defines the internal message structure, and implements parsing/serialising.

Client messages are messages that will be sent to the server by the bot, usually within a callback. Server messages are what the bot receives from the server, having parsed them from the textual line sent.

Parsing is done using Parsec, and messages are parsed in a detailed way according to their command.

Message structure

Client message

data CMsg

Message that will be sent by the bot. The structure is fairly barebones because it's not going to be parsed; whatever is sending the message can implement the structure correctly itself.

Constructors

CMsg

Fields

command :: ClientCmd

argsC :: [String]

data ClientCmd

Commands that will be sent by the bot; deliberately named identically to their textual representations, to make serialising easy.

Constructors

NICK

USER

JOIN

PONG

PRIVMSG

Server message

data SMsg

Message that will be sent by the server and received by the client. These are more detailed structures; messages are parsed into easily usable forms for callbacks. The exception is SNumeric, which is generic for all numeric messages and may have any number of arguments.

Constructors

SPing

Fields

srv :: String

SNotice

Fields

from :: Sender

to :: Recipient

text :: String

SPrivmsg

Fields

from :: Sender

to :: Recipient

text :: String

SNumeric

Fields

fromMaybe :: Maybe Sender

n :: Int

args :: [String]

data Sender

Sender field - can be either user or server.

Constructors

SUser

Fields

nick :: String

user :: String

host :: String

SServer

Fields

host :: String

data Recipient

Recipient field - can either be user or channel.

Constructors

RUser String

RChannel String

Parsing and serialising

joinMsg :: CMsg -> String

Convert a message to a string in the IRC format.

readMsg :: String -> Either ParseError SMsg

Parse a message string in the IRC format.

Parsing internals

parseMsg :: Parser SMsg

The main parser for messages.

parseSender :: Parser Sender

Parse the sender field of a message, which will be a :, followed by either a user in the form nick!user@host, or a server in the form host.

parseRecipient :: Parser Recipient

Parse a recipient field, which might either be a channel in the form #channel-name, or a user in the form nick.

parseArgs :: Parser [String]

Parse the arguments at the end of the message.

parseUntil :: String -> Parser String

Parse until one of the characters in s, or eof.

parseWord :: Parser String

Parse until one of space : r n eof, and eat a trailing space if it's there.