Run

Description

The main logic for the bot. Implements the core networking functionality, as well as finding and applying callbacks.

There's a list of callbacks in callbacks, and for each message, we will invoke all functions in there that successfully pattern match the message.

Support for non-blocking calls on Windows is poor, so we implement our own non-blocking, using a listener thread that blocks, and an MVar between that and the main thread. We can do non-blocking reads on the MVar reliably.

Networking

startBot :: Bot ()

Startup. We should already have initialised global variables about the server and nick from the config, and we use those to connect to the server, sending registration info, then starting the listening loop.

listenMain :: MVar String -> Bot ()

The main listening loop: check whether listenH has read a line, and process that if there is one.

listenH :: Handle -> MVar String -> IO ()

Listen on the socket provided, writing lines we get into the MVar.

Message handling

handleLine :: String -> Bot ()

Handle a line - message string - from the server, invoking the appropriate callbacks.

applyCallbacks :: SMsg -> Bot ()

Apply all message callbacks in callbacks which pattern-match the message.

tryApply :: (a -> b) -> a -> Maybe b

Try to apply f to v - if pattern matching on the argument fails, return Nothing. Otherwise return Just (f v). Used to match callbacks.