Scripts.Core

Description

Implements the core IRC-related functionality, like user registration, responding to pings, joining and leaving channels, and so on.

Documentation

callbacks :: [SEvent -> Bot ()]

Event hooks

Connection

connected :: SEvent -> Bot ()

User registration, and set a callback to join channels.

respondTo376 :: SEvent -> Bot ()

Join channels on 376 (end of MOTD), and then unhook this callback.

Channel management

respondToJoin :: SEvent -> Bot ()

If it's the bot that's joined, update the current channels list.

respondToPart :: SEvent -> Bot ()

If it's the bot that's parted, update the current channels list.

respondToKick :: SEvent -> Bot ()

If it's the bot that's been kicked, update the current channels list.

Other

respondToPing :: SEvent -> Bot ()

Respond to a server PING with PONG.