Scripts.Example

Description

Some example functionality for the bot.

Documentation

respondToChanMsg :: SEvent -> Bot ()

Echo all channel messages, keeping count of the number of messages we've echoed. And enact a delayed action a couple of seconds later.

testCounter :: GlobalKey Int

A variable in the bot's store, keeping track of how many messages we've echoed since startup.