Bot

Description

IRC bot monad, carrying around a state and IO. Most IRC-specific logic is implemented in other modules.

The state, GlobalStore, can hold values of arbitrary type; the keys used, GlobalKey, contain information about the type, so that the values can be retrieved from the Dynamic instances that are actually stored.

The keys use strings for uniqueness, but are also unique by type. So two different keys will only point to the same object if they have the same key string and default values of the same type.

The keys also contain a default value, which is returned whenever the lookup fails. This makes it seem as though the store was initialised with the default values for all the keys.

GlobalStore is intended to provide a scripting interface for addon modules, by providing a way for users to specify arbitrary stateful variables, and share them across modules by exporting the keys.

The Bot monad

newtype Bot a

Bot monad, having State GlobalStore and IO.

Constructors

Bot

Fields

getBot :: StateT GlobalStore IO a

runBot :: Bot a -> GlobalStore -> IO a

Extract the IO computation by running the bot on the given initial state.

empty :: GlobalStore

An empty GlobalStore.

Global storage

getGlobal :: Typeable a => GlobalKey a -> Bot a

Get the value at k in the state.

setGlobal :: Typeable a => GlobalKey a -> a -> Bot ()

Set the value at k to v in the state.

Bot-specific keys for global store

To be initialised in config: The key strings here map to the relevant keys in .ini, as "SECTION.key":

serverHostname :: GlobalKey String

IRC server host

serverPort :: GlobalKey String

IRC server port.

botNick :: GlobalKey String

Nick of bot.

botChan :: GlobalKey String

Channel that bot should join.

logDest :: GlobalKey String

Logging output filename.

configKeys :: [GlobalKey String]

A list of the above keys.

To be initialised elsewhere before running (startBot):

socketH :: GlobalKey Handle

Socket handle

callbackList :: GlobalKey [SMsg -> Bot ()]

List of callbacks to apply to messages

Already initialised

timerList :: GlobalKey [(UTCTime, Bot ())]

List of timed callbacks, to be called at the time specified.

Callback control

addTimer :: UTCTime -> Bot () -> Bot ()

Add a timer to run the given computation at the given time.

runInS :: Integral a => a -> Bot () -> Bot ()

Run the given computation in the given number of seconds.

IO

writeMsg :: CMsg -> Bot ()

Write a message out to the server.

Logging

putLog

Arguments

:: String
Log level (e.g., INFO, ERROR, etc.)

-> String
Line

-> Bot ()

Write a line out to the log, prepended with a timestamp.

putLogInfo :: String -> Bot ()

Log with level INFO.

putLogWarning :: String -> Bot ()

Log with level WARNING.

putLogError :: String -> Bot ()

Log with level ERROR.

Global storage implementation

type GlobalStore = HashMap (TypeRep, String) Dynamic

Store for keeping global state of arbitrary type. It has keys of type (TypeRep, String), allowing for uniqueness for keys of the same type and guaranteeing that keys of different types will never collide. It stores values as Dynamic, extracting them using the type information provided by the key.

data GlobalKey a

Key for GlobalStore - a is the type that's stored. GlobalKey a s creates a global key with:

default value a (which also specifies the type of the value being stored)

id string s (which should be unique for keys of the same type)

Constructors

GlobalKey a String

getGlobalFromStore :: Typeable a => GlobalStore -> GlobalKey a -> a

Look up the key in the store, returning the value found if there is one, or the default value contained in the key otherwise.

setGlobalToStore :: Typeable a => GlobalStore -> GlobalKey a -> a -> GlobalStore

Set the given value to the key in the store.

Re-exported for convenience

liftIO :: MonadIO m => forall a. IO a -> m a

liftM :: Monad m => (a1 -> r) -> m a1 -> m r

mapM :: Traversable t => forall m a b. Monad m => (a -> m b) -> t a -> m (t b)