Bot

Description

IRC bot monad, carrying around a state and IO. Most IRC-specific logic is implemented in other modules.

The state, GlobalStore, can hold values of arbitrary type; the keys used, GlobalKey, contain information about the type, so that the values can be retrieved from the Dynamic instances that are actually stored.

The keys use strings for uniqueness, but are also unique by type. So two different keys will only point to the same object if they have the same key string and default values of the same type.

The keys also contain a default value, which is returned whenever the lookup fails. This makes it seem as though the store was initialised with the default values for all the keys.

GlobalStore is intended to provide a scripting interface for addon modules, by providing a way for users to specify arbitrary stateful variables, and share them across modules by exporting the keys.

The Bot monad

newtype Bot a

Bot monad, having State GlobalStore and IO.

Constructors

Bot

Fields

getBot :: StateT GlobalStore IO a

runBot :: Bot a -> GlobalStore -> IO a

Extract the IO computation by running the bot on the given initial state.

Global storage

data PersistentKey a

PersistentKey a sec name creates a key that transparently reads/writes a value to the config file, as a value called name in section sec.

CacheKey a sec name does the same, but also acts as an instance of GlobalKey a (sec ++ "." ++ name); only reading the value from file once, storing the value in the store in memory and reading from that subequently. It does write to the file for every new value, though.

The objects referred to must have Eq, Read and Show instances for this.

Constructors

PersistentKey a String String

CacheKey a String String

getGlobal :: Typeable a => GlobalKey a -> Bot a

Get the value at k in the state.

setGlobal :: Typeable a => GlobalKey a -> a -> Bot ()

Set the value at k to v in the state.

getGlobal' :: (Eq a, Read a, Show a, Typeable a) => PersistentKey a -> Bot a

getGlobal, but for PersistentKey. This is a separate function because the typeclass restrictions are stronger than getGlobal.

setGlobal' :: (Eq a, Read a, Show a, Typeable a) => PersistentKey a -> a -> Bot ()

setGlobal, but for PersistentKey. This is a separate function because the typeclass restrictions are stronger than setGlobal.

Bot-specific keys for global store

To be initialised in config: The key strings here map to the relevant keys in .ini, as "SECTION.key":

serverHostname :: PersistentKey String

IRC server host

serverPort :: PersistentKey Integer

IRC server port.

botNick :: PersistentKey String

Nick of bot.

botChan :: PersistentKey String

Channel that bot should join.

logDest :: PersistentKey String

Logging output filename.

To be initialised elsewhere before running (startBot):

socketH :: GlobalKey Handle

Socket handle

callbackList :: GlobalKey [(CallbackHandle, SEvent -> Bot ())]

List of callbacks to apply to messages

configFile :: GlobalKey String

Config file filename.

Already initialised

timerList :: GlobalKey [(CallbackHandle, UTCTime, Bot ())]

List of timed callbacks, to be called at the time specified.

Callback control

newtype CallbackHandle

A callback handle referring to callbacks; internally a randomly-generated UUID.

Constructors

CallbackHandle UUID

addCallback :: (SEvent -> Bot ()) -> Bot CallbackHandle

Add the given function as a callback for events; return the handle.

removeCallback :: CallbackHandle -> Bot ()

Remove the callback referred to by the handle. Silently fails if there is no such callback.

addTimer :: UTCTime -> Bot () -> Bot CallbackHandle

Add a timer to run the given computation at the given time, returning a handle.

removeTimer :: CallbackHandle -> Bot ()

Remove the timer referred to by the handle. Silently fails if there is no such timer.

runInS :: Integral a => a -> Bot () -> Bot CallbackHandle

Run the given computation in the given number of seconds, returning a handle.

IO

Messaging

writeMsg :: CMsg -> Bot ()

Write a message out to the server.

sendMessage :: Recipient -> String -> Bot ()

Send a textual message to the recipient

joinChannels :: [String] -> Bot ()

Join the list of channels, where each element is a channel name string (e.g., "#channame")

Logging

putLog

Arguments

:: String
Log level (e.g., INFO, ERROR, etc.)

-> String
Line

-> Bot ()

Write a line out to the log, prepended with a timestamp.

putLogInfo :: String -> Bot ()

Log with level INFO.

putLogWarning :: String -> Bot ()

Log with level WARNING.

putLogError :: String -> Bot ()

Log with level ERROR.

Re-exported for convenience

liftIO :: MonadIO m => forall a. IO a -> m a

module Msg

module Store