Bot.Scripting.Core

Description

Implements the core IRC-related functionality, like user registration, responding to pings, joining and leaving channels, and so on.

Documentation

callbacks :: [SEvent -> Bot ()]

Event hooks

Connection

connected :: SEvent -> Bot ()

User registration, and set a callback to join channels.

disconnected :: SEvent -> Bot ()

Remove all the channels from the current channel list.

manageReadyStatus :: SEvent -> Bot ()

Keep track of ready/unready status.

respondToEndOfMOTD :: SEvent -> Bot ()

Join channels on 376 (end of MOTD), and then unhook this callback.

managePingTimer :: SEvent -> Bot ()

Register or unregister the ping timers when we connect/disconnect. Every once in a while we'll send a ping to the server, to see if the connection is still active. We have a timeout timer, pingTimeout, that will force a reconnect if it expires. When we receive ping response, we reset the timer, so that it's only triggered if we don't receive a timely response.

pingTimeoutLen :: PersistentKey Integer

How long we should wait without receiving a ping response before reconnecting to the server. Pings will be sent at intervals of a third of this.

pingTimer :: Bot ()

Send a ping to the server, every pingTimeoutLen/3 seconds.

respondToPong :: SEvent -> Bot ()

Receive a ping reply from the server and reset the timeout timer.

pingTimeout :: Bot ()

Disconnect from the server by any means by killing the listener thread. This will force a disconnect/reconnect in the main event thread.

Queries

respondToNAMES :: SEvent -> Bot ()

NAMES reply - handles both RPL_NAMREPLY (353) and RPL_ENDOFNAMES (366); the requester will be called on 353 and unhooked; or if, e.g., the channel doesn't exist, the server will only send a 366 and the requester will be called on that.

respondToWHOISUSER :: SEvent -> Bot ()

WHOIS reply - like respondToNAMES, this handles RPL_WHOISUSER (311), RPL_WHOISCHANNELS (319) and RPL_ENDOFWHOIS (318). We keep a list of intermediate responses which we update when we get a 311 and 319, and invoke the callbacks on whatever we've stored when a 318 arrives.

respondToWHOWASUSER :: SEvent -> Bot ()

WHOWAS reply - like respondToNAMES, this handles both RPL_WHOWASUSER (314) and RPL_ENDOFWHOWAS (369). Handle and . Unlike WHOIS, we're never going to get a list of channels, so we don't need to store any intermediate callbacks.

namesResponse :: String -> [User] -> Bot ()

Find the requester for the NAMES query, unhook and invoke it.

whoisResponse :: String -> Maybe User -> Bot ()

Find the requester for the WHOIS/WHOWAS query, unhook and invoke it.

Nick management

nickservAddr :: PersistentKey Sender

The address of the NickServ service on the server.

acquiredNick :: SEvent -> Bot ()

On '001', i.e., getting assigned a nick, start listening for a message from NickServ in case we have to identify for it.

nickservIdent :: SEvent -> Bot ()

On the first message from NickServ, whatever it is, identify for our nick.

identPassword :: PersistentKey [Char]

The password to use to identify the nickname with NickServ.

identSuccess :: SEvent -> Bot ()

When we successfully identify for our nick, we want to request op from ChanServ, in case any ops depended on being identified.

requestOp :: Bot ()

Request op in all channels from ChanServ.

Channel management

chanservAddr :: PersistentKey Sender

The address of the ChanServ service on the server.

respondToJoin :: SEvent -> Bot ()

If it's the bot that's joined, update the current channels list. If we just joined the last remaining channel on the autojoin list, signal Ready.

respondToPart :: SEvent -> Bot ()

If it's the bot that's parted, update the current channels list.

respondToKick :: SEvent -> Bot ()

If it's the bot that's been kicked, update the current channels list, and rejoin the channel after a delay. If this was an autojoin channel, signal UnReady.

unableToJoin :: SEvent -> Bot ()

Catch all the cases when we're unable to join a channel.

unableToJoinChan :: Show a => [Char] -> a -> Bot ()

If we're unable to join a channel, we want to record why, and then retry joining in 10 seconds.

opOnJoin :: SEvent -> Bot ()

Try to get op from ChanServ when we join a channel.

Other

respondToPing :: SEvent -> Bot ()

Respond to a server PING with PONG.