Bot.Scripting.Vote

Description

The voting logic.

Documentation

callbacks :: [SEvent -> Bot ()]

Event hooks

voteChan :: PersistentKey Channel

The channel the bot should take votes in.

actionSettings :: PersistentKey [ActionSetting]

Action settings for each action.

onStartup :: SEvent -> Bot ()

Reinstate any outstanding ban callbacks. Also, clear the current vote for dev purposes.

Vote data

data VoteOpt

Vote options.

Constructors

VoteOpt

Fields

actionO :: ActionSetting

targetO :: String

lenO :: Either Char Integer

reasonO :: Maybe String

hostO :: Maybe String

data ActionSetting

Action setting struct - an action is something that can be voted on for the bot to carry out.

Constructors

ActionSetting

Fields

actAS :: String

The name of the action in the !vote command, e.g., "kick".

actNameAS :: String

The name of the action as a verb, e.g., "kicked".

voteLenAS :: Integer

The length of the vote, in seconds.

voteThresholdAS :: Integer

The minimum number of votes that must be cast for the vote to pass.

useHostAS :: Bool

Whether the action requires a hostname on the part of the target.

defaultLenAS :: Integer

The default duration of the action in seconds, if not specified by a '-l' option.

Command handling

processCommand :: SEvent -> Bot ()

Listen to all channel messages, and extract those which are commands for this bot.

parseCommand :: Parser (SEvent -> Bot ())

Parse the top-level command, returning an action to take.

validateVote :: Channel -> String -> Bot (Maybe VoteOpt)

Parse a !vote command, sending the appropriate error messages, and returning the vote options if the command was valid, or Nothing if not.

parseVoteCmd :: Parser (Either String (String, String, Maybe (Either Char Integer), Maybe String))

Parser for a !vote command, returning parsed parameters if successful, or an error message to send to the channel if not.

Voting logic

vote :: SEvent -> Bot ()

!vote - command to start a vote.

findTarget :: VoteOpt -> Bot ()

Look up the target - a whois/whowas for kick/banning, or checking if they're in the channel for a kick.

startVote :: VoteOpt -> Bot ()

Start the given vote.

voteEnd :: Bot ()

End the current vote, taking the action if successful.

data VoteChoice

A yes/no vote.

Constructors

VYes

VNo

castVote :: VoteChoice -> SEvent -> Bot ()

!yes or !no - vote yes or no, checking the user's host to see if they've already voted.

voteInfo :: SEvent -> Bot ()

Send info about the current vote.

Helpers

showCurrVote :: Bot String

Gets info about the current vote, including end time.

formatVote :: VoteOpt -> String

Pretty print a vote.