Msg

Description

Defines the internal messageevent structure, and implements parsingserialising.

Server events are either messages that the bot has received from the server, like a PRIVMSG, or other events like the bot starting up. They're parsed from the textual line received by a parser using Parsec, and for common commands with specific syntax, they're parsed in a detailed way.

Client messages are messages that will be sent to the server by the bot, usually within a callback. Client messages can be sent generically, as a stock CMsg.

Message structure

Event

data SEvent

Events: a message that's been sent by the server and received by the client, or an event initiated on the bot side like startup. The messages have detailed structures; they're parsed into easily usable forms for callbacks. The exception is SNumeric, which is generic for all numeric messages and may have any number of arguments.

Constructors

SPing

Fields

srv :: String

SNotice

Fields

from :: Sender

to :: Recipient

text :: String

SPrivmsg

Fields

from :: Sender

to :: Recipient

text :: String

SNumeric

Fields

fromMaybe :: Maybe Sender

n :: Int

args :: [String]

SJoin

Fields

from :: Sender

to :: Recipient

SPart

Fields

from :: Sender

to :: Recipient

SKick

Fields

from :: Sender

to :: Recipient

target :: String

reason :: String

SMode

Fields

from :: Sender

modeTarget :: Recipient

modeChanges :: String

Startup

Connected

data Sender

Sender field - can be either user or server.

Constructors

SUser

Fields

nick :: String

user :: String

host :: String

SServer

Fields

host :: String

data Recipient

Recipient field - can either be user or channel.

Constructors

RUser String

RChannel String

Client message

data CMsg

Message that will be sent by the bot. The structure is fairly barebones because it's not going to be parsed; whatever is sending the message can implement the structure correctly itself, or a helper function in Bot can handle that.

Constructors

CMsg

Fields

command :: ClientCmd

argsC :: [String]

data ClientCmd

Commands that will be sent by the bot; deliberately named identically to their textual representations, to make serialising easy.

Constructors

NICK

USER

JOIN

PONG

PRIVMSG

Sending messages

Parsing and serialising

joinMsg :: CMsg -> String

Convert a message to a string in the IRC format.

readMsg :: String -> Either ParseError SEvent

Parse a message string in the IRC format.

Parsing internals

parseMsg :: Parser SEvent

The main parser for messages.

parseSender :: Parser Sender

Parse the sender field of a message, which will be a :, followed by either a user in the form nick!user@host, or a server in the form host.

parseRecipient :: Parser Recipient

Parse a recipient field, which might either be a channel in the form #channel-name, or a user in the form nick.

parseArgs :: Parser [String]

Parse the arguments at the end of the message.

parseUntil :: String -> Parser String

Parse until one of the characters in s, or eof.

parseWord :: Parser String

Parse until one of space : r n eof, and eat a trailing space if it's there.