Scripts.Core

Description

Implements the core IRC-related functionality, like user registration, responding to pings, joining and leaving channels, and so on.

Startup

startup :: SEvent -> Bot ()

User registration, and set a callback to join channels.

respondTo376 :: SEvent -> Bot ()

Join channels on 376 (end of MOTD), and then unhook this callback.

Other

respondToPing :: SEvent -> Bot ()

Respond to a server PING with PONG.